+++ app/CHANGELOG.md
@@ -1,9 +1,13 @@
 ## CHANGELOG
 
-- 2018-11-10 | v.0.17.3 'Phobos':
+- 2018-11-10 | v.0.18.1 'Phobos':
 
   Fix iPad styles of the detailed page
   Rename Burn Rate to Token Age Consumed
+
+- 2018-09-10 | v.0.18.0 'Phobos':
+
+  Add "Trends" Statistic
 
 - 2018-08-10 | v.0.17.2 'Phobos':